+++ client/package.json
@@ -13,7 +13,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-player": "^2.16.0",
-    "react-router-dom": "^6.26.1",
+    "react-router-dom": "^6.26.2",
     "react-scripts": "5.0.1",
     "react-youtube": "^10.1.0",
     "web-vitals": "^2.1.4"
