--- client/src/components/PointsBalance.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Paper, Typography } from '@mui/material';
-
-const PointsBalance = ({ points }) => {
-  return (
-    <Paper elevation={2} sx={{ padding: 2, marginBottom: 2 }}>
-      <Typography variant="h6">Points Balance: {points} NIL</Typography>
-    </Paper>
-  );
-};
-
-export default PointsBalance;
